[maintenance] split experimental settings to a dedicated panel | #BAZEL-1863
